Map file path caching

+++ AngeliA Framework/src/Framework/General/FrameworkUtil.cs
@@ -167,13 +167,9 @@ public static class FrameworkUtil {
 
 	// Misc
 	public static void DeleteAllEmptyMaps (string mapRoot) {
-		var world = new World();
 		foreach (var path in Util.EnumerateFiles(mapRoot, false, $"*.{AngePath.MAP_FILE_EXT}")) {
 			try {
-				if (!world.LoadFromDisk(path)) continue;
-				if (world.EmptyCheck()) {
-					Util.DeleteFile(path);
-				}
+				if (Util.IsExistingFileEmpty(path)) Util.DeleteFile(path);
 			} catch (System.Exception ex) { Util.LogException(ex); }
 		}
 	}
